requirements.txt
boto3
aws-cdk-lib